Cleanup.

+++ test/test-client.js
@@ -9,6 +9,8 @@ var sourceId = 'tmsource://'+basePath+'/test/fixtures-localsource';
 var testPath = path.resolve(path.join(__dirname, '..'));
 var tmpdb = path.join(require('os').tmpdir(), 'tm2-test-' + (+new Date) + '.db');
 var fork = require('child_process').fork;
+var mochabin = path.resolve(path.join(__dirname, '..', 'node_modules', 'mocha-phantomjs', 'bin', 'mocha-phantomjs'));
+var phantombin = require('phantomjs').path;
 
 // Make a working copy of the test database that is excluded in .gitignore
 fs.writeFileSync(tmpdb, fs.readFileSync(path.join(__dirname, 'fixtures-oauth', 'test.db')));
@@ -22,27 +24,25 @@ process.argv.push('--mapboxtile=http://localhost:3001/v4');
 
 require('../index.js');
 
-setTimeout(function() {
-    var mochabin = path.resolve(path.join(__dirname, '..', 'node_modules', 'mocha-phantomjs', 'bin', 'mocha-phantomjs'));
-    var phantombin = require('phantomjs').path;
-    var exit = 0;
-    var child = fork('./node_modules/.bin/mocha-phantomjs', ['http://localhost:3001/style?id='+styleId+'&test=true', '--path='+phantombin]);
-    child.on('exit', function(code) {
+var exit = 0;
+var tests = [
+    'http://localhost:3001/style?id='+styleId+'&test=true',
+    'http://localhost:3001/source?id='+sourceId+'&test=true'
+];
+
+function runTest() {
+    if (!tests.length) {
         fs.unlinkSync(tmpdb);
-        process.exit(code);
+        process.exit(exit);
+        return;
+    }
+    var url = tests.shift();
+    var child = fork(mochabin, [url, '--path='+phantombin]);
+    child.on('exit', function(code) {
+        exit = exit || code;
+        runTest();
     });
-}, 1000);
+}
 
-/*
-# Run mock oauth server and tm2
-node index.js --test --cwd=$testPath --mapboxauth="http://localhost:3001" --mapboxtile="http://localhost:3001/v4" --port=3001 --db="$tmpdb" &
-sleep 2
+setTimeout(runTest, 1000);
 
-./node_modules/.bin/mocha-phantomjs "http://localhost:3001/style?id=$styleid&test=true"
-./node_modules/.bin/mocha-phantomjs "http://localhost:3001/source?id=$sourceid&test=true"
-
-# Remove working database
-rm $tmpdb
-
-exit 0
-*/
